server/api/views.py
from rest_framework import generics
from core.models import Category
from api.serializers import CategorySerializer


class CategoryList(generics.ListAPIView):
    queryset = Category.objects.all()
    serializer_class = CategorySerializer
