server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "Backend side of turkish-literature-calendar"
authors = ["DrShahinstein <1taylanozturk@protonmail.com>"]
license = "GNU General Public License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
